+++ private/get.js
@@ -4,6 +4,6 @@
 */
 module.exports = {
   settings : require( process.cwd() + `/private/${process.env.ENV||'prod'}/settings.js`),
-  svcAcct : process.cwd() + `/private/${process.env.ENV||'prod'}/serviceAccountCredentials.json`
+  svcAcct : require(process.cwd() + `/private/${process.env.ENV||'prod'}/serviceAccountCredentials.json`)
 };